requirements.txt
beautifulsoup4==4.11.1
gspread==5.6.2
oauth2client==4.1.3
pyTelegramBotAPI==4.7.1
python-dotenv==0.21.0
schedule==1.1.0
